requirements.txt
playwright>=1.40.0
playwright-stealth>=2.0.3
flask>=3.0.0
psycopg2-binary>=2.9.9
